views.py refactoring: methods were moved from deleted files (ben.py, utils.py). Question 3 doesn't work now.
