--- DataParis/utils.py
@@ -1,13 +1,0 @@
-from io import BytesIO
-import matplotlib.pyplot as plt
-import base64
-
-def get_graph():
-    buffer = BytesIO()
-    plt.savefig(buffer, format='png')
-    buffer.seek(0)
-    image_png = buffer.getvalue()
-    graph = base64.b64encode(image_png)
-    graph = graph.decode('utf-8')
-    buffer.close()
-    return graph
